Split IDE setup helpers
