+++ cli/python/base_setup/ide.py
@@ -1,22 +1,30 @@
 from __future__ import annotations
 
-import json
-import os
-import subprocess
-import sys
-import tempfile
-from dataclasses import dataclass
-from pathlib import Path
-
 import base_cli
 from base_cli.config import UserConfig
-from base_cli.ide_schema import IDE_DEFINITIONS
-from base_cli.ide_schema import IdeDefinition
+from base_cli.ide_schema import IDE_DEFINITIONS  # pylint: disable=unused-import
+from base_cli.ide_schema import IdeDefinition  # pylint: disable=unused-import
 
-from . import artifacts
-from . import process
 from .checks import ArtifactCheck
-from .errors import ArtifactError
+# pylint: disable=unused-import
+from .ide_diagnostics import IdeDiagnosticSnapshot  # pylint: disable=unused-import
+from .ide_extensions import check_ide_extension  # pylint: disable=unused-import
+from .ide_extensions import check_ide_extensions
+from .ide_extensions import list_ide_extensions  # pylint: disable=unused-import
+from .ide_extensions import reconcile_ide_extensions
+from .ide_installs import check_ide_install  # pylint: disable=unused-import
+from .ide_installs import check_ide_installs
+from .ide_installs import reconcile_ide_install  # pylint: disable=unused-import
+from .ide_installs import reconcile_ide_installs
+from .ide_settings import check_ide_setting  # pylint: disable=unused-import
+from .ide_settings import check_ide_settings
+from .ide_settings import ide_settings_file  # pylint: disable=unused-import
+from .ide_settings import merge_ide_settings  # pylint: disable=unused-import
+from .ide_settings import read_ide_settings  # pylint: disable=unused-import
+from .ide_settings import reconcile_ide_settings
+from .ide_settings import resolve_ide_settings  # pylint: disable=unused-import
+from .ide_settings import write_json_atomic  # pylint: disable=unused-import
+# pylint: enable=unused-import
 from .manifest import BaseManifest, IdeConfig
 
 
@@ -114,416 +122,3 @@ def log_ide_preference_warnings(ctx: base_cli.Context, checks: list[ArtifactChec
         ctx.log.warning(check.message)
         if check.fix:
             ctx.log.warning("Fix: %s", check.fix)
-
-
-@dataclass
-class IdeDiagnosticSnapshot:
-    definition: IdeDefinition
-    _cli_available: bool | None = None
-    _installed_extensions: set[str] | None = None
-    _extension_error: ArtifactError | None = None
-    _settings_file: Path | None = None
-    _current_settings: dict[str, object] | None = None
-    _settings_error: ArtifactError | None = None
-
-    def cli_available(self) -> bool:
-        if self._cli_available is None:
-            self._cli_available = process.command_exists(self.definition.cli)
-        return self._cli_available
-
-    def installed_extensions(self) -> set[str]:
-        if self._installed_extensions is None and self._extension_error is None:
-            try:
-                self._installed_extensions = list_ide_extensions(self.definition)
-            except ArtifactError as exc:
-                self._extension_error = exc
-        if self._extension_error is not None:
-            raise self._extension_error
-        if self._installed_extensions is None:
-            raise RuntimeError(f"{self.definition.label} installed extensions snapshot is unavailable.")
-        return self._installed_extensions
-
-    def settings_file(self) -> Path:
-        if self._settings_file is None:
-            self._settings_file = ide_settings_file(self.definition)
-        return self._settings_file
-
-    def current_settings(self) -> dict[str, object]:
-        if self._current_settings is None and self._settings_error is None:
-            try:
-                self._current_settings = read_ide_settings(self.definition)
-            except ArtifactError as exc:
-                self._settings_error = exc
-        if self._settings_error is not None:
-            raise self._settings_error
-        if self._current_settings is None:
-            raise RuntimeError(f"{self.definition.label} settings snapshot is unavailable.")
-        return self._current_settings
-
-
-def reconcile_ide_installs(ctx: base_cli.Context, manifest: BaseManifest, dry_run: bool) -> None:
-    for ide_name, ide_config in manifest.ide.items():
-        definition = IDE_DEFINITIONS[ide_name]
-        if not ide_config.install:
-            ctx.log.debug("IDE '%s' does not request installation; skipping cask install.", ide_name)
-            continue
-        reconcile_ide_install(ctx, definition, dry_run=dry_run)
-
-
-def reconcile_ide_install(ctx: base_cli.Context, definition: IdeDefinition, dry_run: bool) -> None:
-    command = ["brew", "install", "--cask", definition.cask]
-    if dry_run:
-        process.dry_run_command(ctx, command)
-        return
-
-    if not process.command_exists("brew"):
-        raise ArtifactError(f"Homebrew is required to install {definition.label}.")
-
-    if process.run_check(["brew", "list", "--cask", definition.cask]):
-        ctx.log.info("%s is already installed via Homebrew cask '%s'.", definition.label, definition.cask)
-    else:
-        ctx.log.info("Installing %s via Homebrew cask '%s'.", definition.label, definition.cask)
-        process.run_command(ctx, command)
-
-    if process.command_exists(definition.cli):
-        ctx.log.info("%s CLI '%s' is available on PATH.", definition.label, definition.cli)
-    else:
-        ctx.log.warning(
-            "%s is installed, but CLI '%s' is not on PATH. Enable the IDE shell command before extension setup.",
-            definition.label,
-            definition.cli,
-        )
-
-
-def check_ide_installs(manifest: BaseManifest) -> list[ArtifactCheck]:
-    checks: list[ArtifactCheck] = []
-    for ide_name, ide_config in manifest.ide.items():
-        definition = IDE_DEFINITIONS[ide_name]
-        if ide_config.install:
-            checks.append(check_ide_install(manifest.project_name, definition))
-    return checks
-
-
-def reconcile_ide_extensions(ctx: base_cli.Context, manifest: BaseManifest, dry_run: bool) -> None:
-    for ide_name, ide_config in manifest.ide.items():
-        if not ide_config.extensions:
-            continue
-        definition = IDE_DEFINITIONS[ide_name]
-        if dry_run:
-            for extension in ide_config.extensions:
-                process.dry_run_command(ctx, [definition.cli, "--install-extension", extension])
-            continue
-        if not process.command_exists(definition.cli):
-            ctx.log.warning(
-                "%s CLI '%s' is not on PATH; skipping extension setup.",
-                definition.label,
-                definition.cli,
-            )
-            continue
-        installed_extensions = list_ide_extensions(definition)
-        for extension in ide_config.extensions:
-            if extension in installed_extensions:
-                ctx.log.debug(
-                    "%s extension '%s' is already installed.",
-                    definition.label,
-                    extension,
-                )
-                continue
-            ctx.log.info("Installing %s extension '%s'.", definition.label, extension)
-            process.run_command(ctx, [definition.cli, "--install-extension", extension])
-
-
-def list_ide_extensions(definition: IdeDefinition) -> set[str]:
-    try:
-        completed = process.run_capture(
-            [definition.cli, "--list-extensions"],
-            timeout_seconds=process.DIAGNOSTIC_TIMEOUT_SECONDS,
-        )
-    except subprocess.TimeoutExpired as exc:
-        raise ArtifactError(
-            f"Unable to list {definition.label} extensions with "
-            f"'{definition.cli} --list-extensions': timed out after "
-            f"{process.DIAGNOSTIC_TIMEOUT_SECONDS} seconds."
-        ) from exc
-    if completed.returncode:
-        stderr = (completed.stderr or "").strip()
-        message = f"Unable to list {definition.label} extensions with '{definition.cli} --list-extensions'."
-        if stderr:
-            message = f"{message}\n{stderr}"
-        raise ArtifactError(message)
-    return {line.strip() for line in completed.stdout.splitlines() if line.strip()}
-
-
-def check_ide_extensions(manifest: BaseManifest) -> list[ArtifactCheck]:
-    checks: list[ArtifactCheck] = []
-    for ide_name, ide_config in manifest.ide.items():
-        if not ide_config.extensions:
-            continue
-        definition = IDE_DEFINITIONS[ide_name]
-        snapshot = IdeDiagnosticSnapshot(definition)
-        checks.extend(
-            check_ide_extension(manifest.project_name, definition, extension, snapshot=snapshot)
-            for extension in ide_config.extensions
-        )
-    return checks
-
-
-def check_ide_extension(
-    project: str,
-    definition: IdeDefinition,
-    extension: str,
-    snapshot: IdeDiagnosticSnapshot | None = None,
-) -> ArtifactCheck:
-    snapshot = snapshot or IdeDiagnosticSnapshot(definition)
-    if not snapshot.cli_available():
-        return ArtifactCheck(
-            name=extension,
-            ok=False,
-            message=(
-                f"Cannot check {definition.label} extension '{extension}' "
-                f"because CLI '{definition.cli}' is not on PATH."
-            ),
-            fix=(
-                f"Enable the '{definition.cli}' shell command from {definition.label}, "
-                f"then run 'basectl setup {project}'."
-            ),
-            finding_id="BASE-P110",
-        )
-
-    try:
-        installed_extensions = snapshot.installed_extensions()
-    except ArtifactError as exc:
-        return ArtifactCheck(
-            name=extension,
-            ok=False,
-            message=str(exc),
-            fix=f"basectl setup {project}",
-            finding_id="BASE-P111",
-        )
-
-    if extension in installed_extensions:
-        return ArtifactCheck(
-            name=extension,
-            ok=True,
-            message=f"{definition.label} extension '{extension}' is installed.",
-            fix="",
-            finding_id="BASE-P112",
-        )
-    return ArtifactCheck(
-        name=extension,
-        ok=False,
-        message=f"{definition.label} extension '{extension}' is not installed.",
-        fix=f"basectl setup {project}",
-        finding_id="BASE-P112",
-    )
-
-
-def reconcile_ide_settings(ctx: base_cli.Context, manifest: BaseManifest, dry_run: bool) -> None:
-    for ide_name, ide_config in manifest.ide.items():
-        if not ide_config.settings:
-            continue
-        definition = IDE_DEFINITIONS[ide_name]
-        resolved_settings = resolve_ide_settings(manifest.project_name, ide_config.settings)
-        merge_ide_settings(ctx, definition, resolved_settings, dry_run=dry_run)
-
-
-def resolve_ide_settings(project: str, settings: dict[str, object]) -> dict[str, object]:
-    resolved: dict[str, object] = {}
-    for key, value in settings.items():
-        if key == "python.defaultInterpreterPath" and value == "auto":
-            resolved[key] = str(artifacts.project_venv_dir(project) / "bin" / "python")
-        else:
-            resolved[key] = value
-    return resolved
-
-
-def ide_settings_file(definition: IdeDefinition) -> Path:
-    home = Path(os.environ.get("HOME") or Path.home()).expanduser()
-    if sys.platform == "darwin":
-        return home / "Library" / "Application Support" / definition.settings_app_dir / "User" / "settings.json"
-    config_home = Path(os.environ.get("XDG_CONFIG_HOME") or home / ".config").expanduser()
-    return config_home / definition.settings_app_dir / "User" / "settings.json"
-
-
-def read_ide_settings(definition: IdeDefinition) -> dict[str, object]:
-    settings_file = ide_settings_file(definition)
-    if not settings_file.exists():
-        return {}
-    try:
-        data = json.loads(settings_file.read_text(encoding="utf-8"))
-    except json.JSONDecodeError as exc:
-        raise ArtifactError(f"{settings_file}: invalid JSON: {exc}") from exc
-    if not isinstance(data, dict):
-        raise ArtifactError(f"{settings_file}: expected a JSON object.")
-    return data
-
-
-def merge_ide_settings(
-    ctx: base_cli.Context,
-    definition: IdeDefinition,
-    desired_settings: dict[str, object],
-    dry_run: bool,
-) -> None:
-    settings_file = ide_settings_file(definition)
-    current_settings = read_ide_settings(definition)
-    merged_settings = dict(current_settings)
-    added: dict[str, object] = {}
-
-    for key, value in desired_settings.items():
-        if key not in current_settings:
-            merged_settings[key] = value
-            added[key] = value
-        elif current_settings[key] != value:
-            ctx.log.info(
-                "%s setting '%s' already set by user; leaving intact.",
-                definition.label,
-                key,
-            )
-
-    if not added:
-        ctx.log.debug("%s user settings already contain all Base-managed keys.", definition.label)
-        return
-
-    if dry_run:
-        for key, value in added.items():
-            ctx.log.info(
-                "[DRY-RUN] Would set %s user setting '%s' to %s.",
-                definition.label,
-                key,
-                json.dumps(value, sort_keys=True),
-            )
-        return
-
-    settings_file.parent.mkdir(parents=True, exist_ok=True)
-    write_json_atomic(settings_file, merged_settings)
-    ctx.log.info("Updated %s user settings at '%s'.", definition.label, settings_file)
-
-
-def write_json_atomic(path: Path, data: dict[str, object]) -> None:
-    tmp_path = None
-    try:
-        with tempfile.NamedTemporaryFile("w", encoding="utf-8", dir=path.parent, delete=False) as tmp_file:
-            tmp_path = Path(tmp_file.name)
-            json.dump(data, tmp_file, indent=2, sort_keys=True)
-            tmp_file.write("\n")
-        tmp_path.replace(path)
-    finally:
-        if tmp_path is not None and tmp_path.exists():
-            tmp_path.unlink()
-
-
-def check_ide_settings(manifest: BaseManifest) -> list[ArtifactCheck]:
-    checks: list[ArtifactCheck] = []
-    for ide_name, ide_config in manifest.ide.items():
-        if not ide_config.settings:
-            continue
-        definition = IDE_DEFINITIONS[ide_name]
-        snapshot = IdeDiagnosticSnapshot(definition)
-        resolved_settings = resolve_ide_settings(manifest.project_name, ide_config.settings)
-        checks.extend(
-            check_ide_setting(manifest.project_name, definition, key, value, snapshot=snapshot)
-            for key, value in resolved_settings.items()
-        )
-    return checks
-
-
-def check_ide_setting(
-    project: str,
-    definition: IdeDefinition,
-    key: str,
-    expected_value: object,
-    snapshot: IdeDiagnosticSnapshot | None = None,
-) -> ArtifactCheck:
-    snapshot = snapshot or IdeDiagnosticSnapshot(definition)
-    settings_file = snapshot.settings_file()
-    try:
-        current_settings = snapshot.current_settings()
-    except ArtifactError as exc:
-        return ArtifactCheck(
-            name=f"{definition.label} setting: {key}",
-            ok=False,
-            message=str(exc),
-            fix=f"Repair '{settings_file}' and run 'basectl setup {project}'.",
-            finding_id="BASE-P120",
-        )
-
-    if key not in current_settings:
-        return ArtifactCheck(
-            name=f"{definition.label} setting: {key}",
-            ok=False,
-            message=f"{definition.label} setting '{key}' is absent from '{settings_file}'.",
-            fix=f"basectl setup {project}",
-            finding_id="BASE-P121",
-        )
-    if current_settings[key] == expected_value:
-        return ArtifactCheck(
-            name=f"{definition.label} setting: {key}",
-            ok=True,
-            message=f"{definition.label} setting '{key}' matches the Base manifest.",
-            fix="",
-            finding_id="BASE-P122",
-        )
-    return ArtifactCheck(
-        name=f"{definition.label} setting: {key}",
-        ok=False,
-        message=(
-            f"{definition.label} setting '{key}' is set to {json.dumps(current_settings[key], sort_keys=True)}; "
-            f"expected {json.dumps(expected_value, sort_keys=True)}. Base will not overwrite user settings."
-        ),
-        fix=f"Update '{settings_file}' manually or remove the key and run 'basectl setup {project}'.",
-        finding_id="BASE-P123",
-    )
-
-
-def check_ide_install(project: str, definition: IdeDefinition) -> ArtifactCheck:
-    if not process.command_exists("brew"):
-        return ArtifactCheck(
-            name=f"{definition.label} app",
-            ok=False,
-            message=f"Homebrew is required to check {definition.label} installation.",
-            fix="basectl setup",
-            finding_id="BASE-P130",
-        )
-
-    try:
-        cask_installed = process.run_check(
-            ["brew", "list", "--cask", definition.cask],
-            timeout_seconds=process.DIAGNOSTIC_TIMEOUT_SECONDS,
-        )
-    except subprocess.TimeoutExpired:
-        return ArtifactCheck(
-            name=f"{definition.label} app",
-            ok=False,
-            message=(
-                f"Homebrew cask check for {definition.label} timed out after "
-                f"{process.DIAGNOSTIC_TIMEOUT_SECONDS} seconds."
-            ),
-            fix=f"Retry 'basectl doctor {project}' or inspect Homebrew with 'brew doctor'.",
-            status="warn",
-            finding_id="BASE-P131",
-        )
-    cli_available = process.command_exists(definition.cli)
-
-    if cask_installed and cli_available:
-        return ArtifactCheck(
-            name=f"{definition.label} app",
-            ok=True,
-            message=f"{definition.label} is installed and CLI '{definition.cli}' is on PATH.",
-            fix="",
-            finding_id="BASE-P131",
-        )
-    if not cask_installed:
-        return ArtifactCheck(
-            name=f"{definition.label} app",
-            ok=False,
-            message=f"{definition.label} is not installed via Homebrew cask '{definition.cask}'.",
-            fix=f"basectl setup {project}",
-            finding_id="BASE-P131",
-        )
-    return ArtifactCheck(
-        name=f"{definition.label} CLI",
-        ok=False,
-        message=f"{definition.label} is installed, but CLI '{definition.cli}' is not on PATH.",
-        fix=f"Enable the '{definition.cli}' shell command from {definition.label}.",
-        finding_id="BASE-P132",
-    )
